Add visible Google Translate selector
